za: Refactor and improve the search precision of giphy.
Giphy API precision is far from satifactory.  For example, none
of the API endpoints can really find result for 'pussy' (for any
rating).

Therefore, this commit refactors the giphy ('za') module which
switches to parse the '/search/<string>' result directly.

One caveat, the WeChat gif upload limit is 1MiB.  So the best
effort is to use the 'downsampled' gif.